Add "double differences"
Compare each pixel to the average of the pixel next to it and the pixel above it.
Instead of just the pixel next to it.
In the case of Lenna, this pushed us over the edge and made a result that was 99.833% the size of the original PNG.

Remove some broken code from the histogram.
I never got the user interaction to work right.

https://qoiformat.org/benchmark/ Source for new reference image.